Pull out a utility method
